feat: improve backend frontend connection
