+++ guia-turistica-backend/src/server.ts
@@ -1,27 +1,38 @@
 // src/server.ts
-import express from 'express'
-import morgan from 'morgan'
-import cors from 'cors'
-import authRoutes from './routes/auth'
-import guidesRoutes from './routes/guides'
-import bookingsRoutes from './routes/bookings'
-import reviewsRoutes from './routes/reviews'
+// Setup express API server
+import 'dotenv/config';
+import express from 'express';
+import morgan from 'morgan';
+import cors from 'cors';
 
-const app = express()
+import authRoutes from './routes/auth';
+import guidesRoutes from './routes/guides';
+import bookingsRoutes from './routes/bookings';
+import reviewsRoutes from './routes/reviews';
 
-app.use(cors({
-  origin: ['http://localhost:5173'], // tu Vite dev server
-  credentials: false,
-}));
+const app = express();
 
-app.use(morgan('dev'))
-app.use(express.json()) // 👈 obligatorio para que req.body tenga el JSON
+// Allow the frontend to talk with this API
+const allowedOrigin = process.env.CLIENT_URL ?? 'http://localhost:5173';
+app.use(
+  cors({
+    origin: allowedOrigin,
+    credentials: true,
+  })
+);
 
-app.use('/auth', authRoutes)
-app.use('/guides', guidesRoutes)
-app.use('/bookings', bookingsRoutes)
-app.use('/reviews', reviewsRoutes)
+app.use(morgan('dev'));
+app.use(express.json()); // 👈 obligatorio para que req.body tenga el JSON
 
-app.listen(process.env.PORT ?? 3000, () => {
-  console.log(`API running on http://localhost:${process.env.PORT ?? 3000}`)
-})
+// Health endpoint useful during deployment
+app.get('/health', (_, res) => res.json({ ok: true }));
+
+app.use('/auth', authRoutes);
+app.use('/guides', guidesRoutes);
+app.use('/bookings', bookingsRoutes);
+app.use('/reviews', reviewsRoutes);
+
+const PORT = Number(process.env.PORT) || 3000;
+app.listen(PORT, () => {
+  console.log(`API running on http://localhost:${PORT}`);
+});
