+++ guia-turistica-frontend/src/lib/api.ts
@@ -1,7 +1,10 @@
 import axios from 'axios';
 
+// Ensure baseURL always matches backend origin and has no trailing slash
+const baseURL = (import.meta.env.VITE_API_URL ?? 'http://localhost:3000').replace(/\/$/, '');
+
 const api = axios.create({
-  baseURL: import.meta.env.VITE_API_URL || 'http://localhost:3000',
+  baseURL,
 });
 
 api.interceptors.request.use((cfg) => {
